refactor tandem repeat annotation functions
